refactor(invoices): the row's status and its verbs become components
`app/routes/invoices.tsx` sat at 509 lines against a 510 cap. Pure move:
the rendered markup is byte-identical, every class name and message key
is the one that was there, and all 17 co-located tests — which drive the
real DOM through `createRoutesStub` and click the buttons — pass
unchanged.

Two cells, two components, into `app/components/invoices/` beside the
three that already live there.

`InvoiceRowActions` is the one that earns it. That cell holds a small
state machine — a row is either showing its verbs or showing the
payment-method picker, and which verbs exist depends on whether the
invoice is paid and whether an inspection stands behind it. The page
keeps the state (one picker open at a time across the table) and passes
it in; the component owns what each branch renders. `getPayMethods` went
with it, because the picker is its only caller and the thunk exists for a
reason worth keeping next to its use: Paraglide labels must resolve
inside the per-request locale scope, not freeze at import.

`InvoiceStatusCell` takes the pill, `STATUS_TONE` and `methodLabel`.
Small, but it is the whole of one column and it was the only remaining
reason the route imported `Pill` and `PillTone`.

The IA-122 rationale moved to the control it explains. The client column
carried twenty-six lines describing, among other things, why the Action
column looks the way it does — with the Action column now in another
file that comment would have been the first thing to drift. What is left
in the route says what the client cell decides (the name is text, the row
has no click handler) and points at the component for the rest.

Both components declare their own structural prop type, the convention
`PaymentsModal` and `InvoiceAmountCell` already set: "only the fields
this surface reads".

Baseline ENTRY removed, not tightened — 367 lines is back under the
ordinary 400 rule.
